Added Dialogs for SingletonTutorial
