+++ src/main/java/GUI/Tutorials/TutorialWrapper.java
@@ -1,7 +1,7 @@
 package GUI.Tutorials;
 
 import GUI.ClassGenerators.ClassGenerator;
-import GUI.Tutorials.Singleton.SingletonCard;
+import GUI.Tutorials.Singleton.SingletonTutorial;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.DialogWrapper;
 import com.sun.istack.Nullable;
@@ -10,25 +10,30 @@ import javax.swing.*;
 
 public class TutorialWrapper extends DialogWrapper {
     private ClassGenerator classGenForm;
-    private SingletonCard singletonForm;
+    private SingletonTutorial singletonForm;
 
     public TutorialWrapper(@Nullable Project project) {
         super(project);
-        singletonForm = new SingletonCard();
+        singletonForm = new SingletonTutorial();
         init();
         setTitle("Singleton Generator");
     }
     //https://ayusch.com/intellij-plugin-development-tutorial-handling-user-input/
 //
 //    @Override
-//    protected JComponent createSouthPanel(){
-//        return null;
-//    }
+    protected JComponent createSouthPanel(){
+        return null;
+    }
 
     @Nullable
     @Override
     protected JComponent createCenterPanel() {
         return singletonForm.getContent();
+
+    }
+
+    @Override
+    protected void doOKAction() {
 
     }
